+++ Sources/MacIMEKit/Log.swift
@@ -1,27 +1,9 @@
 import Foundation
 
-public enum LogLevels {
-   case debug
-   case info
-   case warn
-   case error
-   public var desc: String {
-      switch self {
-      case .debug:
-         return "DEBUG"
-      case .info:
-         return "INFO "
-      case .warn:
-         return "WARN "
-      case .error:
-         return "ERROR "
-      }
-   }
-}
-
 public enum Log {
    /// Leave log (to stderr)
-   public static func log(_ msg: String, logLevel _: LogLevels = .info) {
+   public static func log(_ msg: String, logLevel: LogLevel = .info) {
+      guard shouldLog(logLevel) else { return }
       let formatter = ISO8601DateFormatter()
       formatter.timeZone = .current
       let timestamp = formatter.string(from: Date())
@@ -29,6 +11,26 @@ public enum Log {
       if let data = logMsg.data(using: .utf8) {
          FileHandle.standardError.write(data) // standardOutput is not appropriate here
       }
+   }
+
+   public static func getLogLevelByString(_ level: String) -> LogLevel {
+      switch level {
+      case "debug":
+         return .debug
+      case "info":
+         return .info
+      case "warn":
+         return .warn
+      case "error":
+         return .error
+      default:
+         IO.err("Invalid log level: \(level) (Choose from debug|info|warn|error)") // TODO: Should be AppError?
+         exit(1)
+      }
+   }
+
+   public static func shouldLog(_ level: LogLevel) -> Bool {
+      return level >= Runtime.logLevel
    }
 
    public static func debug(_ msg: String) {