+++ Sources/macimed/main.swift
@@ -6,6 +6,10 @@ public var args: [String] = ArgsCommon.getCmdArgs()
 do {
    let state = try ArgsIMED.parse(args)
    Log.info("macimed \(Defaults.version) starting...")
+   Log.debug("a test for debug log")
+   Log.info("a test for info log")
+   Log.warn("a test for warn log")
+   Log.error("a test for error log")
    IMED.setState(state)
    DispatchQueue.global().async {
       try? IMED.serve()